requirements/base.txt
Django==3.0.8
django-allauth==0.42.0
django-crispy-forms==1.9.2
mysqlclient==2.0.1
django-mail-templated==2.6.5
django-model-utils==4.0.0
django-extensions==2.2.9
djangorestframework==3.11.0
